refactor(worker-sessions): satisfy structural quality gates
