+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/factory_runtime_test_helpers_test.go
@@ -169,7 +169,7 @@ func (s *fakeWorkerSessionsService) GetObservation(context.Context, workersessio
 }
 
 func (s *fakeWorkerSessionsService) StreamObservations(context.Context, workersessions.StreamObservationsRequest) (workersessions.ObservationSubscription, error) {
-	return nil, nil
+	return workersessions.ObservationSubscription{}, nil
 }
 
 func (s *fakeWorkerSessionsService) ReadTranscript(context.Context, workersessions.ReadTranscriptRequest) (workersessions.ReadTranscriptResult, error) {
