--- pkg/services/factory_runtime/internal/services/orchestration/runtime/recorded_worker_session_observation_test.go
@@ -1,160 +1,0 @@
-package runtime
-
-import (
-	"context"
-	"encoding/json"
-	"testing"
-	"time"
-
-	"github.com/portpowered/infinite-you/internal/testutil/recordingfixtures"
-	platformclock "github.com/portpowered/infinite-you/pkg/platform/clock"
-	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
-	"github.com/portpowered/infinite-you/pkg/services/recordings"
-	"github.com/portpowered/infinite-you/pkg/services/work"
-	workersessions "github.com/portpowered/infinite-you/pkg/services/worker_sessions"
-	"github.com/portpowered/infinite-you/pkg/services/workers"
-)
-
-func TestRecordedWorkerSessionObservation_ListsHistoricalAttemptsInChronologicalOrder(t *testing.T) {
-	base := time.Date(2026, 8, 8, 12, 0, 0, 0, time.UTC)
-	workID := "work-recorded"
-	events := recordedObservationTestEvents(t, base, workID)
-	ledger := &recordingfixtures.ScriptedRuntimeLedger{Events: events}
-	projector := func(_ []interfaces.FactoryEvent, _ int) (interfaces.FactoryWorldState, error) {
-		return interfaces.FactoryWorldState{
-			WorkItemsByID: map[string]work.FactoryWorkItem{
-				workID: {ID: workID},
-			},
-			CompletedDispatches: []interfaces.FactoryWorldDispatchCompletion{
-				{
-					DispatchID:  "dispatch-late",
-					StartedAt:   base.Add(5 * time.Second),
-					CompletedAt: base.Add(7 * time.Second),
-					WorkItemIDs: []string{workID},
-					Result:      interfaces.WorkstationResult{Outcome: string(workers.OutcomeAccepted)},
-				},
-				{
-					DispatchID:  "dispatch-early",
-					StartedAt:   base.Add(time.Second),
-					CompletedAt: base.Add(3 * time.Second),
-					WorkItemIDs: []string{workID},
-					Result:      interfaces.WorkstationResult{Outcome: string(workers.OutcomeAccepted)},
-				},
-			},
-		}, nil
-	}
-
-	service := newRecordedWorkerSessionObservation(
-		nil,
-		ledger,
-		factory.WorldStateProjector(projector),
-		platformclock.NewDeterministic(base, time.Second),
-	)
-	result, err := service.ListObservations(context.Background(), workersessions.ListObservationsRequest{WorkID: workID})
-	if err != nil {
-		t.Fatalf("ListObservations() error = %v", err)
-	}
-	if len(result.Observations) != 2 {
-		t.Fatalf("ListObservations() returned %d observations, want 2", len(result.Observations))
-	}
-	if got := result.Observations[0].WorkerSessionID; got != "worker-early" {
-		t.Fatalf("first Worker Session = %q, want worker-early", got)
-	}
-	if got := result.Observations[1].WorkerSessionID; got != "worker-late" {
-		t.Fatalf("second Worker Session = %q, want worker-late", got)
-	}
-	if result.Observations[0].AttemptID != "dispatch-early" || result.Observations[0].Duration == nil || *result.Observations[0].Duration != 2*time.Second {
-		t.Fatalf("early recorded observation = %#v, want dispatch-early with 2s duration", result.Observations[0])
-	}
-	if result.Observations[1].AttemptID != "dispatch-late" || result.Observations[1].TurnID != "turn-late" {
-		t.Fatalf("late recorded observation = %#v, want dispatch-late/turn-late", result.Observations[1])
-	}
-}
-
-func TestRecordedWorkerSessionObservation_KnownWorkWithoutSessionsIsExplicitlyEmpty(t *testing.T) {
-	workID := "work-without-sessions"
-	event := interfaces.FactoryEvent{
-		Context: interfaces.FactoryEventContext{Tick: 1, WorkIDs: stringSliceForRecordedTest([]string{workID})},
-		Type:    interfaces.FactoryEventTypeWorkRequest,
-	}
-	ledger := &recordingfixtures.ScriptedRuntimeLedger{Events: []interfaces.FactoryEvent{event}}
-	service := newRecordedWorkerSessionObservation(
-		nil,
-		ledger,
-		func(_ []interfaces.FactoryEvent, _ int) (interfaces.FactoryWorldState, error) {
-			return interfaces.FactoryWorldState{WorkItemsByID: map[string]work.FactoryWorkItem{workID: {ID: workID}}}, nil
-		},
-		platformclock.Real{},
-	)
-
-	result, err := service.ListObservations(context.Background(), workersessions.ListObservationsRequest{WorkID: workID})
-	if err != nil {
-		t.Fatalf("ListObservations() error = %v, want explicit empty success", err)
-	}
-	if result.Observations == nil || len(result.Observations) != 0 {
-		t.Fatalf("observations = %#v, want non-nil empty result", result.Observations)
-	}
-}
-
-func recordedObservationTestEvents(t *testing.T, base time.Time, workID string) []interfaces.FactoryEvent {
-	t.Helper()
-	requestPayload, err := json.Marshal(interfaces.DispatchRequestEventPayload{TransitionID: "review"})
-	if err != nil {
-		t.Fatalf("marshal dispatch request: %v", err)
-	}
-	events := []interfaces.FactoryEvent{
-		{
-			Context: interfaces.FactoryEventContext{
-				Tick: 1, Sequence: 1, EventTime: base.Add(time.Second),
-				DispatchID: stringPointerForRecordedTest("dispatch-early"),
-				WorkIDs:    stringSliceForRecordedTest([]string{workID}),
-			},
-			Type:    interfaces.FactoryEventTypeDispatchRequest,
-			Payload: requestPayload,
-		},
-		{
-			Context: interfaces.FactoryEventContext{
-				Tick: 1, Sequence: 2, EventTime: base.Add(time.Second),
-				DispatchID: stringPointerForRecordedTest("dispatch-early"),
-				RequestID:  stringPointerForRecordedTest("turn-early"),
-			},
-			Type:    interfaces.FactoryEventTypeDispatchWorkerSessionAssoc,
-			Payload: mustMarshalRecordedTest(t, interfaces.DispatchWorkerSessionAssociationEventPayload{WorkerSessionID: "worker-early"}),
-		},
-		{
-			Context: interfaces.FactoryEventContext{
-				Tick: 5, Sequence: 1, EventTime: base.Add(5 * time.Second),
-				DispatchID: stringPointerForRecordedTest("dispatch-late"),
-				WorkIDs:    stringSliceForRecordedTest([]string{workID}),
-			},
-			Type:    interfaces.FactoryEventTypeDispatchRequest,
-			Payload: requestPayload,
-		},
-		{
-			Context: interfaces.FactoryEventContext{
-				Tick: 5, Sequence: 2, EventTime: base.Add(5 * time.Second),
-				DispatchID: stringPointerForRecordedTest("dispatch-late"),
-				RequestID:  stringPointerForRecordedTest("turn-late"),
-			},
-			Type:    interfaces.FactoryEventTypeDispatchWorkerSessionAssoc,
-			Payload: mustMarshalRecordedTest(t, interfaces.DispatchWorkerSessionAssociationEventPayload{WorkerSessionID: "worker-late"}),
-		},
-	}
-	return events
-}
-
-func mustMarshalRecordedTest(t *testing.T, value any) []byte {
-	t.Helper()
-	data, err := json.Marshal(value)
-	if err != nil {
-		t.Fatalf("marshal recorded test payload: %v", err)
-	}
-	return data
-}
-
-func stringPointerForRecordedTest(value string) *string { return &value }
-
-func stringSliceForRecordedTest(value []string) *[]string { return &value }
-
-var _ recordings.RuntimeLedger = (*recordingfixtures.ScriptedRuntimeLedger)(nil)
