+++ pkg/transports/cli/root_terminal_policy_test.go
@@ -17,6 +17,7 @@ import (
 	submitcli "github.com/portpowered/infinite-you/pkg/transports/cli/submit"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/terminalpolicy"
 	workcli "github.com/portpowered/infinite-you/pkg/transports/cli/work"
+	workersessionscli "github.com/portpowered/infinite-you/pkg/transports/cli/worker_sessions"
 )
 
 const (
@@ -476,7 +477,6 @@ func TestProductionRootUsesGeneratedWorkFamilyCutover(t *testing.T) {
 		}
 	}
 }
-
 func TestProductionWorkHandlerRegistryExecutesWatch(t *testing.T) {
 	var got workcli.WatchConfig
 	registry, bindings, err := newWorkHandlerRegistry(
@@ -512,5 +512,120 @@ func TestProductionWorkHandlerRegistryExecutesWatch(t *testing.T) {
 	}
 	if stdout.String() != "watched\n" || stderr.Len() != 0 {
 		t.Fatalf("watch output = %q, diagnostics = %q", stdout.String(), stderr.String())
+	}
+}
+
+func TestWorkerSessionsListCommandMapsManifestInputsToOperation(t *testing.T) {
+	var got workersessionscli.ListConfig
+	factory := withTestInjectedPlatformRoles(CommandFactory{
+		ListWorkerSessions: func(config workersessionscli.ListConfig) error {
+			got = config
+			return nil
+		},
+	})
+	root := factory.NewCommand(nil, nil, nil)
+	root.SetOut(io.Discard)
+	root.SetErr(io.Discard)
+	root.SetArgs([]string{
+		"--json", "--server", "http://factory.test:7437",
+		"worker-sessions", "list", "--work-id", "work-1",
+		"--session", "session-1", "--output", "json",
+	})
+
+	if err := root.Execute(); err != nil {
+		t.Fatalf("execute worker-sessions list: %v", err)
+	}
+	if got.WorkID == "" {
+		t.Fatal("worker sessions list operation was not invoked")
+	}
+	if got.WorkID != "work-1" || got.SessionID != "session-1" || got.Server != "http://factory.test:7437" {
+		t.Fatalf("operation config = %#v, want manifest values", got)
+	}
+	if got.OutputFormat != "json" || !got.JSON {
+		t.Fatalf("output config = %#v, want json output", got)
+	}
+}
+
+func TestWorkerSessionsShowCommandMapsManifestInputsToOperation(t *testing.T) {
+	var got workersessionscli.ShowConfig
+	factory := withTestInjectedPlatformRoles(CommandFactory{
+		ShowWorkerSession: func(config workersessionscli.ShowConfig) error {
+			got = config
+			return nil
+		},
+	})
+	root := factory.NewCommand(nil, nil, nil)
+	root.SetOut(io.Discard)
+	root.SetErr(io.Discard)
+	root.SetArgs([]string{
+		"--json", "--server", "http://factory.test:7437",
+		"worker-sessions", "show", "--provider", "codex", "--kind", "session_id", "--id", "provider-session-1",
+		"--session", "session-1", "--output", "json",
+	})
+
+	if err := root.Execute(); err != nil {
+		t.Fatalf("execute worker-sessions show: %v", err)
+	}
+	if got.Provider != "codex" || got.Kind != "session_id" || got.ID != "provider-session-1" || got.SessionID != "session-1" {
+		t.Fatalf("operation config = %#v, want manifest identity values", got)
+	}
+	if got.Server != "http://factory.test:7437" || got.OutputFormat != "json" || !got.JSON {
+		t.Fatalf("output/config = %#v, want server and json values", got)
+	}
+}
+
+func TestWorkerSessionsStreamCommandMapsManifestInputsToOperation(t *testing.T) {
+	var got workersessionscli.StreamConfig
+	factory := withTestInjectedPlatformRoles(CommandFactory{
+		StreamWorkerSession: func(config workersessionscli.StreamConfig) error {
+			got = config
+			return nil
+		},
+	})
+	root := factory.NewCommand(nil, nil, nil)
+	root.SetOut(io.Discard)
+	root.SetErr(io.Discard)
+	root.SetArgs([]string{
+		"--json", "--server", "http://factory.test:7437",
+		"worker-sessions", "stream", "--provider", "codex", "--kind", "session_id", "--id", "provider-session-1",
+		"--session", "session-1", "--output", "json",
+	})
+
+	if err := root.Execute(); err != nil {
+		t.Fatalf("execute worker-sessions stream: %v", err)
+	}
+	if got.Provider != "codex" || got.Kind != "session_id" || got.ID != "provider-session-1" || got.SessionID != "session-1" {
+		t.Fatalf("operation config = %#v, want manifest identity values", got)
+	}
+	if got.Server != "http://factory.test:7437" || got.OutputFormat != "json" || !got.JSON {
+		t.Fatalf("output/config = %#v, want server and json values", got)
+	}
+}
+
+func TestWorkerSessionsReadCommandMapsManifestInputsToOperation(t *testing.T) {
+	var got workersessionscli.ReadConfig
+	factory := withTestInjectedPlatformRoles(CommandFactory{
+		ReadWorkerSession: func(config workersessionscli.ReadConfig) error {
+			got = config
+			return nil
+		},
+	})
+	root := factory.NewCommand(nil, nil, nil)
+	root.SetOut(io.Discard)
+	root.SetErr(io.Discard)
+	root.SetArgs([]string{
+		"--json", "--server", "http://factory.test:7437",
+		"worker-sessions", "read", "--provider", "codex", "--kind", "session_id", "--id", "provider-session-1",
+		"--session", "session-1", "--output", "json",
+	})
+
+	if err := root.Execute(); err != nil {
+		t.Fatalf("execute worker-sessions read: %v", err)
+	}
+	if got.Provider != "codex" || got.Kind != "session_id" || got.ID != "provider-session-1" || got.SessionID != "session-1" {
+		t.Fatalf("operation config = %#v, want manifest identity values", got)
+	}
+	if got.Server != "http://factory.test:7437" || got.OutputFormat != "json" || !got.JSON {
+		t.Fatalf("output/config = %#v, want server and json values", got)
 	}
 }
